fix: have single revalidate path
